+++ exts/omni.sample.reticle/docs/CHANGELOG.md
@@ -3,6 +3,18 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [1.1.0] - 2022-06-22
+### Added
+- Logic for responding to viewport texture width, height, and fill changes.
+
+### Changed
+- Refactored reticle.py to views.py
+- Fixed bug where Viewport Docs was being treated as viewport.
+- Moved Reticle button to bottom right of viewport to not overlap axes decorator.
+
+### Removed
+- All mutli-viewport logic.
+
 ## [1.0.0] - 2022-05-25
 ### Added
 - Initial add of the Sample Viewport Reticle extension.